requirements.txt
python-telegram-bot
python-dotenv
requests
SQLAlchemy
mysql-connector-python
google-generativeai